Created a ModBlockEntitySideManager class that allows you to register what items can be inserted/extracted to/from what slots. The ExtractorBlockEntity holds this class in the sideManager variable.
When checkIsAllowed is called on the ExtractorBlockEntity within the overwritten canInsert and canExtract classes, the block entities block facing direction is passed, along with the side, slot, and stack.GetItem(). Facing is used to map the world directions to the block sides relevant to the block rotation. Side, slot, and item are then checked against the registered allowances and a boolean is returned.

The current registration is strict, meaning that if the side, slot, and item, do not exactly match a registered allowance, they will be denied.

Next Steps:
Create a custom gui for the extractor.
Add fluid handling to the extractor.
Update extractor recipes for more advanced processing mechanics.
Add tag and other advanced handling to ModBlockEntitySideManager registration.
Add spawner crafting recipes.
Add essence stone and bricks of each color as well as slab,stair, and wall variants of each essence block form.
Add essence pouch: an item that can be carried in your inventory, right-clicked with to open, and has a single chest worth of space exclusively for holding essence. (I would like to find a way that essence items will be automatically sent to the essence pouch when picked up if it is in your inventory.)