The extractor now has a custom gui and is "fully" functional. Crying obsidian drains into a normal obsidian and fills the extractor tank by 250. The tank values are meant to be in millibuckets. It will also add to the tank from a filled bucket and remove from the tank with an empty bucket. The tank has a capacity of 64000mb or 64 buckets of obsidian tears.
Textures for tank capacity still need to be added for a visual representation of the tank's contents. Currently, the tank's value is being printed whenever it is updated.

Next Steps:
Add visual texture for fluid capacity.
Add drained obsidian, obsidian cage piece, and crystallized tears items.
Add crystallizer (a machine that uses obsidian tears to create crystallized tears). This will be very similar to the extractor.
Add spawner crafting recipes.
Add energy system to extractor and crystallizer.
Add essence generator. (Will generate energy based on the tier of the essence given.)
Add the capability for a player to shift right-click on an extractor with a bucket of obsidian tears, which would insert 1000mb and return an empty bucket if there is enough room, or an empty bucket, which would extract 1000mb if there was enough and give the player a bucket of obsidian tears.
Add tag and other advanced handling to ModBlockEntitySideManager registration.
Add essence stone and bricks of each color as well as slab,stair, and wall variants of each essence block form.
Add essence pouch: an item that can be carried in your inventory, right-clicked with to open, and has a single chest worth of space exclusively for holding essence. (I would like to find a way that essence items will be automatically sent to the essence pouch when picked up if it is in your inventory.)